Representing action effects as a tuple of functors (forward and backward)
Currently only have a basic 'Set' effect type, but will add more in when i figure it out. May have to make it a shared pointer. This allows for the effects to be reused in the code as well as allows for validation of the effect itself (see comments in the planner)