That took me way longer than it should have (also it's incorrect still)
Not very good at 3d math yet. Seems to work OK though, and that's whats important. Updating the actor rotation to the control rotation in actions can be done instantaneously for now, as it's not that important. Should look into making a 'Curve watcher' component or something so that I can drive animation using curve values and still plug into the action system