Overhauled GOAPAction, Sort of similar to my original implementation but more refined, can create BP subclasses (with no effects/preconditions)
Not fully tested, but the basic idea works so it is 100% the approach I'm going to move forward with. Essentially, the BP subclasses play the part of the Operator. Iteration times should be much faster, and I believe i will have some results in a few days. Will probably make a subdirectory just for action BPs that I can put on the github